Important Announcement
Thank you for your continued support of HoloQuiz

We regret to announce that UserInterface.java will be graduating on 27th October 2025

Since its debut in HoloQuiz on 26th July 2023, it has served a noble purpose of adding peko to the back of every message sent by HoloQuiz, playing an instrumental role in attracting attention from players server-wide and its selfless contributions as an easter egg to daily questions and events. Being a member of HoloQuiz since the very beginning, we understand that this news may be difficult for many, and we are endlessly grateful to UserInterface.java for its remarkable contributions to HoloQuiz.